+++ src/main/java/com/dao/Order_detailsMapper.java
@@ -6,11 +6,11 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface Order_detailsMapper {
-    long countByExample(Order_detailsExample example);
+    Integer countByExample(Order_detailsExample example);
 
     int deleteByExample(Order_detailsExample example);
 
-    int deleteByPrimaryKey(Long orderDetailsId);
+    int deleteByPrimaryKey(Integer orderDetailsId);
 
     int insert(Order_details record);
 
@@ -18,7 +18,7 @@ public interface Order_detailsMapper {
 
     List<Order_details> selectByExample(Order_detailsExample example);
 
-    Order_details selectByPrimaryKey(Long orderDetailsId);
+    Order_details selectByPrimaryKey(Integer orderDetailsId);
 
     int updateByExampleSelective(@Param("record") Order_details record, @Param("example") Order_detailsExample example);
